test(Media): add MediaObjectTest with 39 tests covering CRUD, uploads, rotation, thumbnails, and usage lifecycle
- Added tests/Unit/Pramnos/Media/MediaObjectTest.php with complete docblocks
  - Default property validation
  - createMd5() file hashing
  - Full save/load/update/delete CRUD lifecycle
  - Comma-separated ID handling in load()
  - Session-based userid and date auto-population
  - Error guard bypass with force=true
  - addImage() processing, thumbnails, deleteOriginal, and MD5 deduplication
  - uploadFile() / uploadImage() with valid/invalid MIME and extension checks
  - _validateUploadFileInput() with missing keys and $_FILES resolution
  - fixStaticPath() path remapping
  - addStringToFilename() suffix insertion
  - rotateLeft() / rotateRight() dimension swapping
  - getThumb() / getMedium() / get() thumbnail resolution and caching
  - addUsage() / loadByUsage() / saveUsage() / removeUsage() lifecycle
  - clearUsage() and multipleUsageUpdate() bulk operations
  - staticGetUsages() filtering by module/specific and duplicate removal
  - getMediaUsages() and getUsages() instance method delegation

fix(Media): patch three source code bugs found during testing
  - MediaObject::saveUsage() had spurious 'update' arg causing SQL syntax error
  - MediaObject::saveUsage() read $_SESSION['uid'] without isset() guard
  - ResizeTools::fastimagecopyresampled() missing (int) casts on GD function
    args caused PHP 8.5 implicit float-to-int deprecations

+++ src/Pramnos/Media/ResizeTools.php
@@ -589,16 +589,16 @@ class ResizeTools extends \Pramnos\Framework\Base
         if (
                 $quality < 5 && (($dst_w * $quality) < $src_w || ($dst_h * $quality) < $src_h)
         ) {
-            $temp = imagecreatetruecolor($dst_w * $quality + 1,
-                    $dst_h * $quality + 1);
+            $temp = imagecreatetruecolor((int)($dst_w * $quality + 1),
+                    (int)($dst_h * $quality + 1));
             imagecopyresized($temp, $src_image, 0, 0, $src_x, $src_y,
-                    $dst_w * $quality + 1, $dst_h * $quality + 1, $src_w, $src_h);
+                    (int)($dst_w * $quality + 1), (int)($dst_h * $quality + 1), $src_w, $src_h);
             imagecopyresampled($dst_image, $temp, $dst_x, $dst_y, 0, 0, $dst_w,
-                    $dst_h, $dst_w * $quality, $dst_h * $quality);
+                    $dst_h, (int)($dst_w * $quality), (int)($dst_h * $quality));
             unset($temp);
         } else {
-            imagecopyresampled($dst_image, $src_image, $dst_x, $dst_y, $src_x,
-                    $src_y, $dst_w, $dst_h, $src_w, $src_h);
+            imagecopyresampled($dst_image, $src_image, (int)$dst_x, (int)$dst_y, (int)$src_x,
+                    (int)$src_y, (int)$dst_w, (int)$dst_h, (int)$src_w, (int)$src_h);
         }
         return true;
     }
